+++ profile_dfu/src/main/java/no/nordicsemi/android/dfu/profile/main/view/DFUScreen.kt
@@ -38,7 +38,7 @@ import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Settings
+import androidx.compose.material.icons.outlined.Settings
 import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
@@ -105,7 +105,7 @@ private fun DFUAppBar(onEvent: (DFUViewEvent) -> Unit) {
             }
             IconButton(onClick = { onEvent(OnSettingsButtonClick) }) {
                 Icon(
-                    imageVector = Icons.Default.Settings,
+                    imageVector = Icons.Outlined.Settings,
                     contentDescription = stringResource(id = R.string.dfu_settings_action)
                 )
             }
